requirements.txt
matplotlib==3.10.0
numpy==2.2.1
pytest_benchmark==5.1.0
